requirements.txt
pick==0.6.4
pywin32==224
psutil==5.6.1
windows-curses==1.1
progressbar2==3.39.3
